+++ app-factory.js
@@ -1,7 +1,9 @@
+const Component = require('./component')
+
 module.exports = (renderFactory) => (mainComponent, container) => {
   let mainComponentInstance
   const render = renderFactory(() => mainComponentInstance.view(), container)
-  mainComponentInstance = mainComponent.create(render)
+  mainComponentInstance = Component(render, mainComponent)
   render()
   return mainComponentInstance
 }