+++ h.js
@@ -1,7 +1,5 @@
 const {h} = require('ultradom')
 
-module.exports = (name, attributes, ...children) => {
-  return typeof name === 'function'
-    ? name(attributes, children)
-    : h(name, attributes, children)
-}
+module.exports = (name, attributes, ...children) => typeof name === 'function'
+  ? name(attributes, children)
+  : h(name, attributes, children)