+++ index.js
@@ -1,15 +1,8 @@
 const {render} = require('ultradom')
-
 const h = require('./h')
-const Component = require('./component')
-const App = require('./app')
-
-function renderFactory (view, container) {
-  return () => requestAnimationFrame(() => render(view(), container))
-}
+const AppFactory = require('./app-factory')
 
 module.exports = {
   h,
-  Component,
-  App: App(renderFactory)
+  App: AppFactory((view, container) => () => requestAnimationFrame(() => render(view(), container)))
 }