+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0 (June 1, 2018)
+ * replaced Component with plain old object
+ * added $global and $parent references to state
+ * added ES5 iife build under dist/
+ * removed assertions around view function, now just crashes
+
 ## 2.0.0 (April 28, 2018)
  * fixed a bug where children where not passed to child components
  * added specification test